requirements.txt
matplotlib
numpy
pillow
opencv-python
pyvista
scikit-learn